ui design improved

+++ app/src/main/java/com/example/sleepmonitor/settings.kt
@@ -5,8 +5,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "settings")
 data class Settings(
     @PrimaryKey val id: Int = 1,  // We'll use a single entry, so we set a fixed ID
-    var threshold: Double,
+    var threshold: Int,
     var mobileNumber: String,
     var alarmDuration: Int
 )
-
